File Browser and Folder refactor. alpha)
